feat(web): add more static routes
