+++ apps/og-image/package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "@vercel/og": "^1.0.0",
-    "next": "^13.5.1",
+    "next": "^13.5.4",
     "react": "latest",
     "react-dom": "latest",
     "zod": "^3.21.4"
