+++ apps/web/package.json
@@ -53,8 +53,8 @@
     "lz-string": "^1.5.0",
     "million": "2.6.0-beta.12",
     "monaco-vim": "^0.4.0",
-    "next": "^13.5.1",
-    "next-international": "1.0.0-rc.6",
+    "next": "^13.5.4",
+    "next-international": "1.1.2",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-day-picker": "^8.8.0",
