+++ apps/web/src/app/[locale]/page.tsx
@@ -4,16 +4,11 @@ import { Community } from './_components/community/community';
 import { Features } from './_components/features';
 import { Hero } from './_components/hero';
 import { WaitlistBanner } from './_components/waitlist-banner';
-import { getStaticParams } from '~/locales/server';
 import { setStaticParamsLocale } from 'next-international/server';
 import { buildMetaForDefault } from '../metadata';
 
 export async function generateMetadata(): Promise<Metadata> {
   return buildMetaForDefault({});
-}
-
-export function generateStaticParams() {
-  return getStaticParams();
 }
 
 export default async function Index({ params: { locale } }: { params: { locale: string } }) {
