+++ apps/web/src/app/[locale]/providers.tsx
@@ -8,17 +8,17 @@ import { ThemeProvider } from 'next-themes';
 import React, { Suspense } from 'react';
 import { Toolbar } from '~/components/toolbar';
 import { I18nProviderClient } from '~/locales/client';
-import en from '~/locales/en';
-import { FeatureFlagProvider } from './feature-flag-provider';
+import { FeatureFlagProvider } from '../feature-flag-provider';
 import type { Corner } from '@tanstack/react-query-devtools/build/lib/utils';
 
 interface Props {
+  locale: string
   children: React.ReactNode;
 }
 
 const queryClient = new QueryClient();
 
-export function Providers({ children }: Props) {
+export function Providers({ locale, children }: Props) {
   const position: Corner = 'top-right';
   return (
     <QueryClientProvider client={queryClient}>
@@ -27,7 +27,7 @@ export function Providers({ children }: Props) {
         <ThemeProvider attribute="class">
           <FeatureFlagProvider>
             <TooltipProvider>
-              <I18nProviderClient fallbackLocale={en}>{children}</I18nProviderClient>
+              <I18nProviderClient locale={locale}>{children}</I18nProviderClient>
             </TooltipProvider>
           </FeatureFlagProvider>
           <Suspense>
